# src/main/resources/static/openapi/openapi.yml
openapi: 3.0.3
info:
  title: Hello World API
  description: A simple Hello World API with Spring Boot and OpenAPI
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Local development server
tags:
  - name: hello-world-controller
    description: Hello World Controller
  - name: calculator-controller
    description: Calculator Controller
paths:
  /api/hello:
    get:
      tags:
        - hello-world-controller
      summary: Get a hello message
      description: Returns a greeting message with the provided name or a default greeting
      operationId: sayHello
      parameters:
        - name: name
          in: query
          description: Name to be greeted
          required: false
          schema:
            type: string
            default: World
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HelloResponse'
  /api/info:
    get:
      tags:
        - hello-world-controller
      summary: Get application info
      description: Returns information about the application configuration
      operationId: getInfo
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InfoResponse'
  /api/calculator/add:
    get:
      tags:
        - calculator-controller
      summary: Add two numbers using SOAP service
      description: Calls a SOAP web service to add two numbers
      operationId: add
      parameters:
        - name: a
          in: query
          description: First number
          required: true
          schema:
            type: integer
            format: int32
        - name: b
          in: query
          description: Second number
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CalculatorResponse'
  /api/calculator/subtract:
    get:
      tags:
        - calculator-controller
      summary: Subtract two numbers using SOAP service
      description: Calls a SOAP web service to subtract second number from first number
      operationId: subtract
      parameters:
        - name: a
          in: query
          description: First number
          required: true
          schema:
            type: integer
            format: int32
        - name: b
          in: query
          description: Second number
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CalculatorResponse'
components:
  schemas:
    HelloResponse:
      type: object
      properties:
        message:
          type: string
          example: "Hello, World! Welcome to Spring Boot with OpenAPI and Spring!"
    InfoResponse:
      type: object
      properties:
        version:
          type: string
          example: "3.2.3"
        javaVersion:
          type: string
          example: "21"
        description:
          type: string
          example: "Spring Boot application with OpenAPI documentation"
    CalculatorResponse:
      type: object
      properties:
        operation:
          type: string
          example: "add"
        a:
          type: integer
          format: int32
          example: 10
        b:
          type: integer
          format: int32
          example: 5
        result:
          type: integer
          format: int32
          example: 15
        requestId:
          type: string
          example: "123e4567-e89b-12d3-a456-426614174000"